uses json instead of array params
