+++ packages/payload/src/admin/views/list.ts
@@ -42,8 +42,14 @@ export type ListViewClientProps = {
   disableBulkEdit?: boolean
   enableRowSelections?: boolean
   hasCreatePermission: boolean
+  /**
+   * @deprecated
+   */
   listPreferences?: ListPreferences
   newDocumentURL: string
+  /**
+   * @deprecated
+   */
   preferenceKey?: string
   renderedFilters?: Map<string, React.ReactNode>
   resolvedFilterOptions?: Map<string, ResolvedFilterOptions>
