+++ packages/ui/src/providers/ListQuery/context.ts
@@ -1,0 +1,7 @@
+import { createContext, useContext } from 'react'
+
+import type { IListQueryContext } from './types.js'
+
+export const ListQueryContext = createContext({} as IListQueryContext)
+
+export const useListQuery = (): IListQueryContext => useContext(ListQueryContext)
