+++ packages/ui/src/providers/ListQuery/index.tsx
@@ -1,48 +1,19 @@
 'use client'
-import type { ColumnPreference, ListQuery, PaginatedDocs, Sort, Where } from 'payload'
+import type { ColumnPreference, ListQuery, Where } from 'payload'
 
 import { useRouter, useSearchParams } from 'next/navigation.js'
 import { isNumber } from 'payload/shared'
 import * as qs from 'qs-esm'
-import React, { createContext, useCallback, useContext, useEffect, useMemo, useState } from 'react'
+import React, { useCallback, useEffect, useMemo, useState } from 'react'
+
+import type { ListQueryProps } from './types.js'
 
 import { useListDrawerContext } from '../../elements/ListDrawer/Provider.js'
 import { useRouteTransition } from '../../providers/RouteTransition/index.js'
 import { parseSearchParams } from '../../utilities/parseSearchParams.js'
+import { ListQueryContext } from './context.js'
 
-type ContextHandlers = {
-  handlePageChange?: (page: number) => Promise<void>
-  handlePerPageChange?: (limit: number) => Promise<void>
-  handleSearchChange?: (search: string) => Promise<void>
-  handleSortChange?: (sort: string) => Promise<void>
-  handleWhereChange?: (where: Where) => Promise<void>
-}
-
-export type ListQueryProps = {
-  readonly children: React.ReactNode
-  readonly collectionSlug?: string
-  readonly columns: ColumnPreference[]
-  readonly data: PaginatedDocs
-  readonly defaultLimit?: number
-  readonly defaultSort?: Sort
-  readonly modifySearchParams?: boolean
-  readonly onQueryChange?: (query: ListQuery) => void
-  readonly preferenceKey?: string
-}
-
-export type ListQueryContext = {
-  data: PaginatedDocs
-  defaultLimit?: number
-  defaultSort?: Sort
-  query: ListQuery
-  refineListData: (args: ListQuery) => Promise<void>
-} & ContextHandlers
-
-const Context = createContext({} as ListQueryContext)
-
-type ColumnFromQuery = { accessor: string; active: 'false' | 'true' }
-
-export const useListQuery = (): ListQueryContext => useContext(Context)
+export { useListQuery } from './context.js'
 
 export const ListQueryProvider: React.FC<ListQueryProps> = ({
   children,
@@ -59,21 +30,14 @@ export const ListQueryProvider: React.FC<ListQueryProps> = ({
   const { startRouteTransition } = useRouteTransition()
 
   const searchParams = useMemo<ListQuery>(() => {
-    const parsed: ListQuery = Object.entries(parseSearchParams(rawSearchParams)).reduce(
-      (acc, [key, value]) => {
-        if (key === 'columns') {
-          acc[key] = (value as ColumnFromQuery[]).map((c: ColumnFromQuery) => ({
-            accessor: c.accessor,
-            active: c.active === 'true',
-          }))
-        }
+    const parsed = parseSearchParams(rawSearchParams)
+    const result: ListQuery = parsed
 
-        return acc
-      },
-      {},
-    )
+    if (parsed.columns) {
+      result.columns = JSON.parse(parsed.columns as string) as ColumnPreference[]
+    }
 
-    return parsed
+    return result
   }, [rawSearchParams])
 
   const { onQueryChange } = useListDrawerContext()
@@ -115,7 +79,12 @@ export const ListQueryProvider: React.FC<ListQueryProps> = ({
 
       if (modifySearchParams) {
         startRouteTransition(() =>
-          router.replace(`${qs.stringify(newQuery, { addQueryPrefix: true })}`),
+          router.replace(
+            `${qs.stringify(
+              { ...newQuery, columns: JSON.stringify(newQuery.columns) },
+              { addQueryPrefix: true },
+            )}`,
+          ),
         )
       } else if (
         typeof onQueryChange === 'function' ||
@@ -203,13 +172,17 @@ export const ListQueryProvider: React.FC<ListQueryProps> = ({
       if (shouldUpdateQueryString) {
         setCurrentQuery(newQuery)
         // Do not use router.replace here to avoid re-rendering on initial load
-        window.history.replaceState(null, '', `?${qs.stringify(newQuery)}`)
+        window.history.replaceState(
+          null,
+          '',
+          `?${qs.stringify({ ...newQuery, columns: JSON.stringify(newQuery.columns) })}`,
+        )
       }
     }
   }, [defaultSort, defaultLimit, router, modifySearchParams, columns])
 
   return (
-    <Context.Provider
+    <ListQueryContext.Provider
       value={{
         data,
         handlePageChange,
@@ -222,6 +195,6 @@ export const ListQueryProvider: React.FC<ListQueryProps> = ({
       }}
     >
       {children}
-    </Context.Provider>
+    </ListQueryContext.Provider>
   )
 }
