+++ packages/ui/src/providers/ListQuery/types.ts
@@ -1,0 +1,29 @@
+import type { ColumnPreference, ListQuery, PaginatedDocs, Sort, Where } from 'payload'
+
+type ContextHandlers = {
+  handlePageChange?: (page: number) => Promise<void>
+  handlePerPageChange?: (limit: number) => Promise<void>
+  handleSearchChange?: (search: string) => Promise<void>
+  handleSortChange?: (sort: string) => Promise<void>
+  handleWhereChange?: (where: Where) => Promise<void>
+}
+
+export type ListQueryProps = {
+  readonly children: React.ReactNode
+  readonly collectionSlug?: string
+  readonly columns: ColumnPreference[]
+  readonly data: PaginatedDocs
+  readonly defaultLimit?: number
+  readonly defaultSort?: Sort
+  readonly modifySearchParams?: boolean
+  readonly onQueryChange?: (query: ListQuery) => void
+  readonly preferenceKey?: string
+}
+
+export type IListQueryContext = {
+  data: PaginatedDocs
+  defaultLimit?: number
+  defaultSort?: Sort
+  query: ListQuery
+  refineListData: (args: ListQuery) => Promise<void>
+} & ContextHandlers
